+++ data/src/main/kotlin/com/woowla/ghd/data/remote/RemoteDataSourceImpl.kt
@@ -4,16 +4,19 @@ import com.apollographql.apollo3.ApolloClient
 import com.apollographql.apollo3.api.ApolloResponse
 import com.apollographql.apollo3.api.Operation
 import com.apollographql.apollo3.network.http.HttpInfo
+import com.woowla.ghd.AppLogger
 import com.woowla.ghd.data.remote.entities.ApiGhdRelease
 import com.woowla.ghd.data.remote.mappers.toGhdRelease
 import com.woowla.ghd.domain.entities.ApiResponse
 import com.woowla.ghd.data.remote.mappers.toPullRequest
 import com.woowla.ghd.data.remote.mappers.toRelease
+import com.woowla.ghd.data.remote.mappers.toRepository
 import com.woowla.ghd.domain.entities.GhdRelease
 import com.woowla.ghd.domain.entities.PullRequestWithRepoAndReviews
 import com.woowla.ghd.domain.entities.RateLimit
 import com.woowla.ghd.domain.entities.ReleaseWithRepo
 import com.woowla.ghd.domain.entities.RepoToCheck
+import com.woowla.ghd.domain.entities.Repository
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.engine.cio.CIO
@@ -31,6 +34,7 @@ class RemoteDataSourceImpl(
     private val ghRepo: String,
     private val apolloClient: ApolloClient,
     private val ktorClient: HttpClient,
+    private val appLogger: AppLogger,
 ) : RemoteDataSource {
     companion object {
         fun apolloClientInstance(authorizationInterceptor: AuthorizationInterceptor) = ApolloClient.Builder()
@@ -68,6 +72,36 @@ class RemoteDataSourceImpl(
 
             ApiResponse(
                 data = data.map { it.toPullRequest(repoToCheck) },
+                rateLimit = rateLimit
+            )
+        }
+    }
+
+    override suspend fun search(
+        text: String?,
+        owner: String?,
+    ): Result<ApiResponse<List<Repository>>> {
+        return runCatching {
+            val searchQueryItems = buildList {
+                if (!owner.isNullOrBlank()) {
+                    add("owner:$owner")
+                }
+            }
+
+            val searchQuery = SearchRepositoryQuery(
+                query = "$text " + searchQueryItems.joinToString(separator = "+"),
+                first = 25
+            )
+            val searchResponse = apolloClient.query(searchQuery).execute()
+            val data = searchResponse
+                .dataAssertNoErrors
+                .search
+                .edges
+                ?.mapNotNull { it?.node?.onRepository } ?: listOf()
+            val rateLimit = searchResponse.getHeadersAsMap().getRateLimit()
+
+            ApiResponse(
+                data = data.map { it.toRepository() },
                 rateLimit = rateLimit
             )
         }
@@ -116,6 +150,7 @@ class RemoteDataSourceImpl(
             ApiResponse(data = data.toGhdRelease(), rateLimit = rateLimit)
         }
     }
+
 
     private fun HttpResponse.getHeadersAsMap(): Map<String, String> {
         return headers.entries().associate { it.key to it.value.joinToString(separator = ", ") }