+++ domain-api/src/main/kotlin/com/woowla/ghd/domain/services/RepositoryService.kt
@@ -1,0 +1,7 @@
+package com.woowla.ghd.domain.services
+
+import com.woowla.ghd.domain.entities.Repository
+
+interface RepositoryService {
+    suspend fun search(text: String? = null, user: String? = null): Result<List<Repository>>
+}