+++ ui/src/main/kotlin/com/woowla/ghd/presentation/components/Cards.kt
@@ -29,6 +29,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.PointerEventType
 import androidx.compose.ui.input.pointer.onPointerEvent
 import androidx.compose.ui.text.style.TextOverflow
@@ -43,6 +44,7 @@ import com.woowla.compose.icon.collections.tabler.tabler.outline.Edit
 import com.woowla.compose.icon.collections.tabler.tabler.outline.Filter
 import com.woowla.compose.icon.collections.tabler.tabler.outline.Refresh
 import com.woowla.compose.icon.collections.tabler.tabler.outline.Rocket
+import com.woowla.compose.icon.collections.tabler.tabler.outline.Star
 import com.woowla.compose.icon.collections.tabler.tabler.outline.Tag
 import com.woowla.compose.icon.collections.tabler.tabler.outline.Trash
 import com.woowla.ghd.domain.entities.CommitCheckRollupStatus
@@ -50,6 +52,7 @@ import com.woowla.ghd.domain.entities.PullRequestStateExtended
 import com.woowla.ghd.domain.entities.PullRequestWithRepoAndReviews
 import com.woowla.ghd.domain.entities.ReleaseWithRepo
 import com.woowla.ghd.domain.entities.RepoToCheck
+import com.woowla.ghd.domain.entities.Repository
 import com.woowla.ghd.domain.entities.SyncResultEntryWithRepo
 import com.woowla.ghd.domain.entities.SyncResultWithEntriesAndRepos
 import com.woowla.ghd.domain.entities.anyCommentedOrChangesRequested
@@ -63,6 +66,54 @@ import com.woowla.ghd.presentation.decorators.RepoToCheckDecorator
 import com.woowla.ghd.presentation.decorators.SyncResultDecorator
 import com.woowla.ghd.presentation.decorators.SyncResultEntryDecorator
 import com.woowla.ghd.utils.openWebpage
+
+@OptIn(ExperimentalLayoutApi::class)
+@Composable
+fun RepositoryCard(
+    repository: Repository,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    CardListItem(
+        modifier = modifier,
+        onClick = onClick,
+        title = repository.fullName,
+        subtitle = repository.description,
+        leadingContent = {
+            Avatar(
+                imageUrl = repository.imageUrl,
+            )
+        },
+        supportingContent = {
+            FlowRow(
+                horizontalArrangement = Arrangement.spacedBy(8.dp),
+                verticalArrangement = Arrangement.spacedBy(4.dp)
+            ) {
+                Tag(
+                    text = repository.stargazerCount.toString(),
+                    icon = Tabler.Outline.Star,
+                    color = "stars".toColor(),
+                )
+                val primaryLanguageName = repository.primaryLanguageName
+                if (!primaryLanguageName.isNullOrBlank()) {
+                    Tag(
+                        text = primaryLanguageName,
+                        icon = null,
+                        color = "primaryLanguageName".toColor(),
+                    )
+                }
+                val licenseInfo = repository.licenseInfo
+                if (!licenseInfo.isNullOrBlank()) {
+                    Tag(
+                        text = licenseInfo,
+                        icon = null,
+                        color = "licenseInfo".toColor(),
+                    )
+                }
+            }
+        }
+    )
+}
 
 @Composable
 fun SynResultCard(